[CS-3824]: Move and improve InputAmount
