+++ cardstack/src/components/Input/__tests__/InputAmount.test.tsx
@@ -2,7 +2,7 @@ import { NativeCurrency } from '@cardstack/cardpay-sdk';
 import React from 'react';
 
 import { render } from '../../../test-utils';
-import { InputAmount, CURRENCY_DISPLAY_MODE } from '../InputAmount';
+import { CURRENCY_DISPLAY_MODE, InputAmount } from '../InputAmount/InputAmount';
 
 jest.mock('@rainbow-me/navigation/', () => ({
   useNavigation: jest.fn(() => ({
@@ -18,7 +18,8 @@ describe('InputAmount', () => {
 
     const props = {
       currencyDisplayMode: CURRENCY_DISPLAY_MODE.NO_DISPLAY,
-      nativeCurrency: nativeCurrency,
+      selectedCurrency: nativeCurrency,
+      onCurrencyChange: jest.fn(),
       inputValue,
       setInputValue,
     };
